# Stockwave-Postman-Collection-Json2OpenApi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Stockwave",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "variables": {}
    }
  ],
  "paths": {
...
    "/certificado": {
      "get": {
        "tags": [
          "Certificado"
        ],
        "summary": "Listar Certificados",
        "operationId": "ListarCertificados",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "Certificado"
        ],
        "summary": "Cadastrar Certificado",
        "operationId": "CadastrarCertificado",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CadastrarCertificadoRequest"
                  },
                  {
                    "example": {
                      "dt_certificado": "1997-07-01"
                    }
                  }
                ]
              },
              "example": {
                "dt_certificado": "1997-07-01"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/certificado/1": {
      "get": {
        "tags": [
          "Certificado"
        ],
        "summary": "Exibir Certificados",
        "operationId": "ExibirCertificados",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      },
      "put": {
        "tags": [
          "Certificado"
        ],
        "summary": "Atualizar Certificado",
        "operationId": "AtualizarCertificado",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/AtualizarCertificadoRequest"
                  },
                  {
                    "example": {
                      "dt_certificado": "2000-10-22",
                      "id_certificado": 1
                    }
                  }
                ]
              },
              "example": {
                "dt_certificado": "2000-10-22",
                "id_certificado": 1
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
    "/produto": {
      "get": {
        "tags": [
          "Produto"
        ],
        "summary": "Listar Produtos",
        "operationId": "ListarProdutos",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "Produto"
        ],
        "summary": "Cadastrar Produto",
        "operationId": "CadastrarProduto",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CadastrarProdutoRequest"
                  },
                  {
                    "example": {
                      "imagem_produto": "HTTP://WWW.IMAGEM.COM/IMAGEM03",
                      "nome_produto": "PRODUTO 07",
                      "qtd_produto": 10,
                      "valor_produto": 50.0
                    }
                  }
                ]
              },
              "example": {
                "imagem_produto": "HTTP://WWW.IMAGEM.COM/IMAGEM03",
                "nome_produto": "PRODUTO 07",
                "qtd_produto": 10,
                "valor_produto": 50.0
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/produto/2": {
      "get": {
        "tags": [
          "Produto"
        ],
        "summary": "Exibir Produto",
        "operationId": "ExibirProduto",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      },
      "put": {
        "tags": [
          "Produto"
        ],
        "summary": "Atualizar Produto",
        "operationId": "AtualizarProduto",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/AtualizarProdutoRequest"
                  },
                  {
                    "example": {
                      "id_produto": 2,
                      "imagem_produto": "HTTP://WWW.IMAGEM.COM/IMAGEM03",
                      "nome_produto": "PRODUTO 03",
                      "qtd_produto": 10,
                      "valor_produto": 50.0
                    }
                  }
                ]
              },
              "example": {
                "id_produto": 2,
                "imagem_produto": "HTTP://WWW.IMAGEM.COM/IMAGEM03",
                "nome_produto": "PRODUTO 03",
                "qtd_produto": 10,
                "valor_produto": 50.0
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AtualizarCertificadoRequest": {
        "title": "AtualizarCertificadoRequest",
        "required": [
          "dt_certificado",
          "id_certificado"
        ],
        "type": "object",
        "properties": {
          "dt_certificado": {
            "type": "string"
          },
          "id_certificado": {
            "type": "integer",
            "format": "int32"
          }
        },
        "example": {
          "dt_certificado": "2000-10-22",
          "id_certificado": 1
        }
      },
      "CadastrarCertificadoRequest": {
        "title": "CadastrarCertificadoRequest",
        "required": [
          "dt_certificado"
        ],
        "type": "object",
        "properties": {
          "dt_certificado": {
            "type": "string"
          }
        },
        "example": {
          "dt_certificado": "1997-07-01"
        }
      },
...
      "AtualizarProdutoRequest": {
        "title": "AtualizarProdutoRequest",
        "required": [
          "id_produto",
          "imagem_produto",
          "nome_produto",
          "qtd_produto",
          "valor_produto"
        ],
        "type": "object",
        "properties": {
          "id_produto": {
            "type": "integer",
            "format": "int32"
          },
          "imagem_produto": {
            "type": "string"
          },
          "nome_produto": {
            "type": "string"
          },
          "qtd_produto": {
            "type": "integer",
            "format": "int32"
          },
          "valor_produto": {
            "type": "number"
          }
        },
        "example": {
          "id_produto": 2,
          "imagem_produto": "HTTP://WWW.IMAGEM.COM/IMAGEM03",
          "nome_produto": "PRODUTO 03",
          "qtd_produto": 10,
          "valor_produto": 50.0
        }
      },
      "CadastrarProdutoRequest": {
        "title": "CadastrarProdutoRequest",
        "required": [
          "imagem_produto",
          "nome_produto",
          "qtd_produto",
          "valor_produto"
        ],
        "type": "object",
        "properties": {
          "imagem_produto": {
            "type": "string"
          },
          "nome_produto": {
            "type": "string"
          },
          "qtd_produto": {
            "type": "integer",
            "format": "int32"
          },
          "valor_produto": {
            "type": "number"
          }
        },
        "example": {
          "imagem_produto": "HTTP://WWW.IMAGEM.COM/IMAGEM03",
          "nome_produto": "PRODUTO 07",
          "qtd_produto": 10,
          "valor_produto": 50.0
        }
      },
...
    }
  },
...
}